Create README.md
This is a simple yet very useful solution for parking problems in urban areas. The customer will use the app to find all the empty parking spaces around their locality and choose accordingly. Wondering where the parking spaces are coming from? . Its simple really all the people who own a piece or a huge land that can be used as a parking space will register their spaces in this app . These places can range from as less as an apartment parking till a whole empty land parking. Due to this the customer will find a place to park their vehicles and the land owners can get a passive income aswell. This app can play a major role in reducing traffic in urban areas and a easy money making method